requirements.txt
py-cord==2.5.0
Pillow==10.0.0
Pillow==10.3.0
python-dotenv==1.0.1
Requests==2.31.0
websocket_client==1.8.0
